Move errors and error handlers into their organized directories. Then initialize error handlers from create_app().
