+++ app/forms/email_form.py
@@ -3,6 +3,6 @@ from wtforms import StringField
 from wtforms.validators import DataRequired, Email
 
 
-class EmailPasswordForm(FlaskForm):
+class MPVEmailForm(FlaskForm):
     email = StringField('Email', validators=[DataRequired(), Email()])
     units = StringField('Units', validators=[DataRequired()])
